requirements.txt
pytesseract==0.3.10
Pillow==10.4.0
reportlab==4.2.0
python-docx==1.1.2
pdf2docx==0.5.8
pandas==2.2.2
openpyxl==3.1.5
phonenumbers==8.13.45
vobject==0.9.6.1
PyYAML==6.0.2
python-dateutil==2.9.0.post0
fastapi==0.115.5
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
